ECO-611: refactor, extract static methods.
