+++ packages/ui/src/containers/DeployContractsContainer.ts
@@ -36,7 +36,7 @@ export type ComplexType =
   | 'List'
   | 'FixedList';
 
-type SupportedType = CLType.SimpleMap[keyof CLType.SimpleMap] | ComplexType;
+export type SupportedType = CLType.SimpleMap[keyof CLType.SimpleMap] | ComplexType;
 
 export const ComplexTypesString = [
   'Bytes',
@@ -404,217 +404,6 @@ export class DeployContractsContainer {
     }
   }
 
-  private static isSimpleType(t: SupportedType) {
-    return Object.values(CLType.Simple).includes(t);
-  }
-
-  private static buildSimpleArgs(
-    firstType: CLType.SimpleMap[keyof CLType.SimpleMap],
-    secondType: KeyType | null,
-    argValueInJson: any,
-    uRefAccessRight?: Key.URef.AccessRightsMap[keyof Key.URef.AccessRightsMap]
-  ) {
-    let clValueInstance;
-    switch (firstType) {
-      case CLType.Simple.BOOL:
-        clValueInstance = Args.Instances.bool(argValueInJson);
-        break;
-      case CLType.Simple.I32:
-        clValueInstance = Args.Instances.i32(argValueInJson);
-        break;
-      case CLType.Simple.I64:
-        clValueInstance = Args.Instances.i64(argValueInJson);
-        break;
-      case CLType.Simple.U8:
-        clValueInstance = Args.Instances.u8(argValueInJson);
-        break;
-      case CLType.Simple.U32:
-        clValueInstance = Args.Instances.u32(argValueInJson);
-        break;
-      case CLType.Simple.U64:
-        clValueInstance = Args.Instances.u64(argValueInJson);
-        break;
-      case CLType.Simple.U128:
-        clValueInstance = Args.Instances.u128(argValueInJson);
-        break;
-      case CLType.Simple.U256:
-        clValueInstance = Args.Instances.u256(argValueInJson);
-        break;
-      case CLType.Simple.U512:
-        clValueInstance = Args.Instances.u512(argValueInJson);
-        break;
-      case CLType.Simple.UNIT:
-        clValueInstance = Args.Instances.unit();
-        break;
-      case CLType.Simple.STRING:
-        clValueInstance = Args.Instances.string(argValueInJson);
-        break;
-      case CLType.Simple.KEY:
-      case CLType.Simple.UREF:
-        clValueInstance = DeployContractsContainer.buildKeyOrUrefInstance(
-          firstType,
-          secondType,
-          argValueInJson,
-          uRefAccessRight
-        );
-        break;
-    }
-    return clValueInstance;
-  }
-
-  private buildArgument(arg: FormState<DeployArgument>) {
-    const argValueStr: string = arg.$.value.value;
-    const type = arg.$.type.$;
-    let clValueInstance: CLValueInstance;
-    const argValueStrInJson = JSON.parse(argValueStr);
-    if (DeployContractsContainer.isSimpleType(type)) {
-      clValueInstance = DeployContractsContainer.buildSimpleArgs(
-        type as CLType.SimpleMap[keyof CLType.SimpleMap],
-        arg.$.secondType.$,
-        argValueStrInJson,
-        arg.$.URefAccessRight.$
-      );
-    } else {
-      switch (type) {
-        case 'Bytes':
-          clValueInstance = Args.Instances.bytes(
-            decodeBase16(argValueStrInJson)
-          );
-          break;
-        case 'Bytes (Fixed Length)':
-          clValueInstance = Args.Instances.bytesFixedLength(
-            decodeBase16(argValueStrInJson)
-          );
-          break;
-        case 'Tuple':
-          clValueInstance = this.buildTupleTypeArg(
-            arg.$.tupleInnerDeployArgs.$,
-            argValueStrInJson
-          );
-          break;
-        case 'Map':
-          clValueInstance = this.buildMapTypeArg(
-            arg.$.mapInnerDeployArgs.$,
-            argValueStrInJson
-          );
-          break;
-        case 'List':
-          clValueInstance = this.buildListTypeArg(
-            arg.$.listInnerDeployArgs.$,
-            argValueStrInJson,
-            false
-          );
-          break;
-        case 'FixedList':
-          clValueInstance = this.buildListTypeArg(
-            arg.$.listInnerDeployArgs.$,
-            argValueStrInJson,
-            true
-          );
-          break;
-        case 0:
-        case 1:
-        case 2:
-        case 3:
-        case 4:
-        case 5:
-        case 6:
-        case 7:
-        case 8:
-        case 9:
-        case 10:
-        case 11:
-        case 12:
-          throw new Error('Failed creating arguments');
-      }
-    }
-
-    const deployArg = new Deploy.Arg();
-    deployArg.setName(arg.$.name.value);
-    deployArg.setValue(clValueInstance);
-    return deployArg;
-  }
-
-  private buildTupleTypeArg(
-    tupleInnerDeployArgs: FormDeployArgument[],
-    argValueStr: any
-  ) {
-    const argsTupleList: CLValueInstance[] = [];
-    tupleInnerDeployArgs.forEach((arg, i) => {
-      const type = arg.$.type.$;
-      if (!DeployContractsContainer.isSimpleType(type)) {
-        throw new Error("Don't support nest types in Tuple instance");
-      }
-      const secondType = arg.$.secondType.$;
-      const uRefAccessRight = arg.$.URefAccessRight.$;
-      argsTupleList.push(
-        DeployContractsContainer.buildSimpleArgs(
-          type as CLType.SimpleMap[keyof CLType.SimpleMap],
-          secondType,
-          argValueStr[i],
-          uRefAccessRight
-        )
-      );
-    });
-    if (argsTupleList.length === 1) {
-      return Args.Instances.tuple1(argsTupleList[0]);
-    } else if (argsTupleList.length === 2) {
-      return Args.Instances.tuple2(argsTupleList[0], argsTupleList[1]);
-    } else {
-      return Args.Instances.tuple3(
-        argsTupleList[0],
-        argsTupleList[1],
-        argsTupleList[2]
-      );
-    }
-  }
-
-  private static buildKeyOrUrefInstance(
-    firstType: number,
-    secondType: KeyType | null,
-    argValueStr: string,
-    uRefAccessRight?: 1 | 4 | 7 | 6 | 5 | 0 | 2 | 3 | undefined
-  ): CLValueInstance {
-    const value = new CLValueInstance.Value();
-    const clType = new CLType();
-    if (firstType === CLType.Simple.KEY) {
-      const key = new Key();
-      let keyType = secondType as KeyType;
-      let valueInByteArray = decodeBase16(argValueStr);
-      switch (keyType) {
-        case KeyType.ADDRESS:
-          const address = new Key.Address();
-          address.setAccount(valueInByteArray);
-          key.setAddress(address);
-          break;
-        case KeyType.HASH:
-          const hash = new Key.Hash();
-          hash.setHash(valueInByteArray);
-          key.setHash(hash);
-          break;
-        case KeyType.UREF:
-          const uRef = new Key.URef();
-          uRef.setUref(valueInByteArray);
-          uRef.setAccessRights(uRefAccessRight!);
-          key.setUref(uRef);
-          break;
-      }
-      value.setKey(key);
-      clType.setSimpleType(CLType.Simple.KEY);
-    } else {
-      const URef = new Key.URef();
-      URef.setAccessRights(uRefAccessRight!);
-      URef.setUref(decodeBase16(argValueStr));
-      value.setUref(URef);
-      clType.setSimpleType(CLType.Simple.UREF);
-    }
-
-    const clValueInstance = new CLValueInstance();
-    clValueInstance.setClType(clType);
-    clValueInstance.setValue(value);
-    return clValueInstance;
-  }
-
   /**
    * Implement validator for DeployArgument which can be used in formstate library,
    * If a truthy string is returned it represents a validation error.
@@ -762,66 +551,5 @@ export class DeployContractsContainer {
       DeployContractsContainer.PersistentKey,
       JSON.stringify(state)
     );
-  }
-
-  private buildListTypeArg(
-    listInnerDeployArgs: FormDeployArgument[],
-    argValueInJson: any,
-    isFixedList: boolean
-  ) {
-    const firstType = listInnerDeployArgs[0].$.type.$;
-    if (!DeployContractsContainer.isSimpleType(firstType)) {
-      throw new Error("Don't support nest types");
-    }
-    const secondType = listInnerDeployArgs[0].$.secondType.$;
-    const uRefAccessRight = listInnerDeployArgs[0].$.URefAccessRight.$;
-    const argsList = argValueInJson.map((arg: any) => {
-      return DeployContractsContainer.buildSimpleArgs(
-        firstType as CLType.SimpleMap[keyof CLType.SimpleMap],
-        secondType,
-        arg,
-        uRefAccessRight
-      );
-    });
-    if (isFixedList) {
-      return Args.Instances.fixedList(argsList);
-    } else {
-      return Args.Instances.list(argsList);
-    }
-  }
-
-  private buildMapTypeArg(
-    mapInnerDeployArgs: FormDeployArgument[],
-    argValueInJson: any
-  ) {
-    const keyType = mapInnerDeployArgs[0].$.type.$;
-    const valueType = mapInnerDeployArgs[1].$.type.$;
-    if (
-      !DeployContractsContainer.isSimpleType(keyType) ||
-      !DeployContractsContainer.isSimpleType(valueType)
-    ) {
-      throw new Error("Don't support nest types in Map instance");
-    }
-    const keySecondType = mapInnerDeployArgs[0].$.secondType.$;
-    const keyURefAccessRight = mapInnerDeployArgs[0].$.URefAccessRight.$;
-
-    const valueSecondType = mapInnerDeployArgs[1].$.secondType.$;
-    const valueURefAccessRight = mapInnerDeployArgs[1].$.URefAccessRight.$;
-    const mapEntries = argValueInJson.map((arg: any) => {
-      const key = DeployContractsContainer.buildSimpleArgs(
-        keyType as CLType.SimpleMap[keyof CLType.SimpleMap],
-        keySecondType,
-        arg.key,
-        keyURefAccessRight
-      );
-      const value = DeployContractsContainer.buildSimpleArgs(
-        valueType as CLType.SimpleMap[keyof CLType.SimpleMap],
-        valueSecondType,
-        arg.value,
-        valueURefAccessRight
-      );
-      return [key, value];
-    });
-    return Args.Instances.map(mapEntries);
   }
 }
